# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/branches": {
            "get": {
                "summary": "Returns all branches",
                "tags": [
                    "Branches"
                ],
                "responses": {
                    "200": {
                        "description": "List of all branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Branch"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new branch",
                "tags": [
                    "Branches"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branches/{id}": {
            "get": {
                "summary": "Get a branch by ID",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "put": {
                "summary": "Update a branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Branch deleted successfully"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "summary": "Returns all orders",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "List of all orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new order",
                "tags": [
                    "Orders"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "summary": "Update an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
        "/api/suppliers": {
            "get": {
                "summary": "Returns all suppliers",
                "tags": [
                    "Suppliers"
                ],
                "responses": {
                    "200": {
                        "description": "List of all suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Supplier"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new supplier",
                "tags": [
                    "Suppliers"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "summary": "Get a supplier by ID",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "put": {
                "summary": "Update a supplier",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a supplier",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supplier deleted successfully"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            }
        },
        "/api/suppliers/{id}/status": {
            "get": {
                "summary": "Get the status of a supplier",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "INACTIVE",
                                                "APPROVED",
                                                "PENDING"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Branch": {
                "type": "object",
                "required": [
                    "branchId",
                    "name",
                    "headquartersId"
                ],
                "properties": {
                    "branchId": {
                        "type": "integer",
                        "description": "The unique identifier for the branch"
                    },
                    "headquartersId": {
                        "type": "integer",
                        "description": "The ID of the headquarters this branch belongs to"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the branch"
                    },
                    "address": {
                        "type": "string",
                        "description": "Physical address of the branch"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the branch"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the branch"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "Name of the primary contact person"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the branch"
                    }
                }
            },
...
            "Order": {
                "type": "object",
                "required": [
                    "orderId",
                    "branchId",
                    "orderDate"
                ],
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "description": "The unique identifier for the order"
                    },
                    "branchId": {
                        "type": "integer",
                        "description": "The ID of the branch that placed the order"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the order was placed"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the order"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the order"
                    },
                    "status": {
                        "type": "string",
                        "description": "The current status of the order",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    }
                }
            },
...
            "Supplier": {
                "type": "object",
                "required": [
                    "supplierId",
                    "name"
                ],
                "properties": {
                    "supplierId": {
                        "type": "integer",
                        "description": "The unique identifier for the supplier"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the supplier"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "Name of the primary contact person"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the supplier"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the supplier"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the supplier"
                    },
                    "active": {
                        "type": "boolean",
                        "description": "Whether the supplier is active"
                    },
                    "verified": {
                        "type": "boolean",
                        "description": "Whether the supplier is verified"
                    }
                }
            }
        }
    },
...
}